requirements.txt
pandas
Faker
cryptography
psycopg2-binary
python-dotenv
